Fix usage of double, use float instead

+++ sblib/src/i2c/SHT2x.cpp
@@ -96,7 +96,7 @@ float SHT2xClass::GetDewPoint(void)
   float temperature = GetTemperature();
 
   // Calculate the intermediate value 'gamma'
-  float gamma = log(humidity / 100) + WATER_VAPOR * temperature / (BAROMETRIC_PRESSURE + temperature);
+  float gamma = logf(humidity / 100) + WATER_VAPOR * temperature / (BAROMETRIC_PRESSURE + temperature);
   // Calculate dew point in Celsius
   float dewPoint = BAROMETRIC_PRESSURE * gamma / (WATER_VAPOR - gamma);
